requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
mysql-connector-python
python-dotenv
pydantic
